fix: display only ASCII graphic blobs as strings
This change changes the display tree labels to be more concervative.

Before, we printed everything that is valid UTF-8 as strings.
This causes problems if the input is a byte string with lots of zeros.

For example, canister id 0x00000000000000070101 gets printed as ""
in my terminal.

After this change, we will only print blobs as strings if all the
bytes are graphic ASCII characters.